requirements.txt
streamlit==1.49.1
tensorflow==2.19.1
opencv-python==4.11.0.86
mediapipe==0.10.21
numpy==1.26.4
pandas==2.3.2
matplotlib==3.10.6
Pillow==10.0.0
gdown==4.7.1